src/requirements.txt
cloudshell-shell-core>=4.0,<4.1
cloudshell-networking>=5.1,<5.2
cloudshell-networking-devices>=2.3,<2.4